Remove dead code in amidst.mojangapi.world.versionfeatures
