+++ src/main/java/amidst/mojangapi/world/versionfeatures/VersionFeatureBuilder.java
@@ -12,9 +12,7 @@ import amidst.mojangapi.minecraftinterface.RecognisedVersion;
 public class VersionFeatureBuilder<V> {
 	private V defaultValue = null;
 	private RecognisedVersion previous = null;
-	private RecognisedVersion previousExact = null;
-	private final List<VersionFeatureEntry<V>> exactMatches = new LinkedList<>();
-	private final List<VersionFeatureEntry<V>> entriesNewestFirst = new LinkedList<>();
+	private final List<VersionFeature.Entry<V>> entriesNewestFirst = new LinkedList<>();
 
 	public VersionFeatureBuilder<V> init(V defaultValue) {
 		if (this.defaultValue == null) {
@@ -22,23 +20,6 @@ public class VersionFeatureBuilder<V> {
 			return this;
 		} else {
 			throw new IllegalStateException("only one default value is allowed");
-		}
-	}
-
-	public VersionFeatureBuilder<V> exact(RecognisedVersion version, V value) {
-		Objects.requireNonNull(version, "version cannot be null");
-		if (this.defaultValue == null) {
-			throw new IllegalStateException("you need to specify a default value first");
-		} else if (previousExact != null && version == previousExact) {
-			throw new IllegalStateException("you can only specify one value per version");
-		} else if (previousExact != null && RecognisedVersion.isOlder(version, previousExact)) {
-			throw new IllegalStateException("you have to specify versions in ascending order");
-		} else if (!entriesNewestFirst.isEmpty()) {
-			throw new IllegalStateException("you have to specify all exact matches before the first since");
-		} else {
-			previousExact = version;
-			exactMatches.add(0, new VersionFeatureEntry<>(version, value));
-			return this;
 		}
 	}
 
@@ -52,7 +33,7 @@ public class VersionFeatureBuilder<V> {
 			throw new IllegalStateException("you have to specify versions in ascending order");
 		} else {
 			previous = version;
-			entriesNewestFirst.add(0, new VersionFeatureEntry<>(version, value));
+			entriesNewestFirst.add(0, new VersionFeature.Entry<>(version, value));
 			return this;
 		}
 	}
@@ -60,7 +41,6 @@ public class VersionFeatureBuilder<V> {
 	public VersionFeature<V> construct() {
 		return new VersionFeature<>(
 				defaultValue,
-				Collections.unmodifiableList(exactMatches),
 				Collections.unmodifiableList(entriesNewestFirst));
 	}
 }
